# docs/api/synect-openapi.yaml
openapi: 3.0.3
info:
  title: Synect API
  version: 1.0.0
  description: OpenAPI contract for core Synect backend endpoints.
servers:
  - url: http://localhost:3001
    description: Local development server
...
paths:
...
  /api/internships:
    get:
      tags: [Internships]
      summary: List internships
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: Paginated internships list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    post:
      tags: [Internships]
      summary: Create internship
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Internship created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiSuccess:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Request successful
        data:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
...